chore: initial move in of Terraform code
This moves the Terraform code from what is now
https://github.com/uktrade/data-workspace-frontend

This is part of a wider tidy-up of code in
https://github.com/uktrade/data-workspace-frontend - separating out code that
different teams are responsible for and/or are released in different ways.

This repo is to be the "core" repository that is the entry point to Data
Workspace, which includes deploying it.