docs: move from MkDocs to Eleventy
This moves the documentation from the MkDocs To Eleventy.

This is primarily because the plugin we used to make the docs GOV.UK-ish
https://github.com/ministryofjustice/mkdocs-tech-docs-template is now
deprecated.

The corresponding Eleventy plugin that we're now using,
https://x-govuk.github.io/govuk-eleventy-plugin/, is active, and the same
plugin we're using in other DBT projects, e.g.

https://stream-zip.docs.trade.gov.uk/
https://stream-unzip.docs.trade.gov.uk/
https://pg-bulk-ingest.docs.trade.gov.uk/
https://stream-read-xbrl.docs.trade.gov.uk/

This also renamed the documentation from "technical" documentation, to
"developer" documentation. This I think makes it clearer that this isn't
documentation for anything technical "inside" Data Workspace