feat: support for a second MLFlow to store artefacts in a bucket
This adds support to the VPC endpoint configuration for a second MLFlow
instance to access its bucket.

It's not great that the number is hard coded to 2 - others parts of the
Terraform support 0 or more. However, leaving that tidy up to further work
since it was already hard coded (to 1), and so not a meaningful step backwards.