feat: tidy up of MWAA names
Mainly opting to not call it "dataflow" because that's fairly internal to DBT.
Going for something more generic so at least in theory everything this repo can
be used elsewhere.

Also making it so the bucket name is a variable that can be set - there is one
namespace for all customers of AWS, so there is a chance of collision whatever
we pick really.