feat: ECS Airflow
This is a basic ECS Airflow set up, comparing it to the MWAA Airflow set up.

GOV.UK PaaS is shutting down, so we need to move off. MWAA Airflow doesn't give
us the ability to run DAGs without access to all other DAGs' credentials, so we
are investigating how straight forward it is to do that with our own Airflow on
ECS.

Co-authored-by: Tash Boyse <57753415+nboyse@users.noreply.github.com>
Co-authored-by: Michal Charemza <michal@charemza.name>
Co-authored-by: Mohizur Khan <mohizurkhan@digital.trade.gov.uk>
Co-authored-by: Josh Wong <166488409+joshwong-cs@users.noreply.github.com>