feat: per-team role arn prefix available to Airflow scheduler
Our (slightly) custom ECS executor for Airflow will run tasks using a different
role for each team. The arn of the roles will follow a pattern - a prefix
followed by the folder that the dag of the Airflow task is in.

This change makes that prefix available to the Airflow scheduler that runs the
executor, so the executor can then choose the correct role to run each task
under.