feat: tightening security groups between scheduler and webserver
Currently scheduler and webserver share the same secuirty groups, meaning load
server can connect to the scheduler when it should not be able to.

Therefore, we are spliting the security groups, one for scheduler and one for
webserver.

This introduces a concept that we haven't used so far: "prefix lists" on a VPC
endpoint. This means the scheduler doesn't need to talk to the world in order
to pull its docker images from S3, and can just communicate with S3.

We should probably tighten up the permissions on the S3 endpoint, but leaving
that to another change (because this change doesn't allow anything that wasn't
previously allowed).