feat: add ArangoDB
(Optionally) adds Arango to Data Workspace

Includes Arango ECS service connecting to the main ECS cluster.  ArangoDB
instance run using EC2 to allow for Elastic Block Storage to attach to ArangoDB
instances when spun up (auto-scaling group). There is a load balancer to
connect the Data workspace tools to the ArangoDB instance.

The EBS volume mounts to /data directory within the instance, and the ArangoDB
container task definition mounts the instance's /data directory to
/var/lib/arangodb3 directory within the container (as this is where data is
stored in an Arango database).

As the Arango EC2 instance is in a private subnet and the Datasets VPC does not
have access to the internet, VPC Endpoints were used too allow the EC2
insrtance to connect to ECS. The endpoints were placed in the Datasets VPC
(which did require DNS hostnames to be enabled) and policies were created and
attach to the endpoints to restrict access to only the Arango instance.

Co-authored-by: Isobel Daley <71375510+isobel-daley-6point6@users.noreply.github.com>
Co-authored-by: Zayn-6point6 <zayn.rassam@6point6.co.uk>
Co-authored-by: Michal Charemza <michal@charemza.name>
Co-authored-by: Ares Galamatis <argyris.galamatis@digital.trade.gov.uk>
Co-authored-by: Peter Woodcock <peter.woodcock@digital.trade.gov.uk>
Co-authored-by: Mohizur Khan <mohizurkhan@digital.trade.gov.uk>