feat: buckets for so-called legacy Airflow pipelines
This creates a bucket per team for Airflow pipelines, the so-called legacy
pipelines that ingest into S3. This is part of the project of moving our
Airflow instance out of GOV.UK PaaS which is shutting down, doing essentially a
lift and shift into Data Workspace proper.

At the moment getting credentials for the bucket to the team code is slightly
faffy using Terraform's "output" and then copy/pasting them into secrets
manager, but it's the easiest way for now. Using roles would be "better" but
would require a bit of a code change in data-flow, so leaving that to further
work so we don't get blocked on it and can still migrate pipelines.

Co-authored-by: Mohizur Khan <mohizurkhan@digital.trade.gov.uk>
Co-authored-by: Peter Woodcock <peter.woodcock@digital.trade.gov.uk>
Co-authored-by: Michal Charemza <michal@charemza.name>