feat: use role-based perms with temporary creds for data-flow bucket access
This follows on from 8bb3fa871c6b1e9c858a24feaecb5c9d1e7d259f, changing the
permanent users with long lasting credentials for data-flow to access buckets
to a ECS-role based solution (that under the hood uses temporary credentials).

This is both for security, and for the fact that we already use role-based
perms for logging, and having _both_ long lasting creds via AWS_* environment
variables and role based permissions in the same ECS task instance is tricky to
work with - better one or the other, and so choosing the more secure one.