feat: separate webserver and dags in terms of security groups (and allow egress on 5432)
The DAGs now run using the dag-processor security group, which while maybe
slightly oddly named, is the one that is meant for "team controlled" code.

This also allows egress from dags on port 5432 (PosgreSQL) to everything,
similar to how egress on 443 (HTTPS) is allowed to everything for pipelines to
connect to external sources.

This change also removes the webserver's ability to connect to GitHub - it
shouldn't need to do this.