feat: put the Airflow webserver on a fixed public IP address
This is maybe a bit odd, but it helps certain VPNs route requests
appropriately, especially in the case of Airflow being on the subdomain of Data
Workspace, which has a wildcard TLS certificate and can cause problems when
users flip between Airflow and the rest of Data Workspace (and specifically
browsers re-using the connection of one for another).

This also meant moving to a network load balancer, since application load
balancers apparently don't support fixed IP addresses.