feat: add lifecycle policies to all ECR repos, except visualisation-base images
Because we only used tagged images in ECS, to reduce costs and to avoid alerts
for vulnrabilities that have since been addressed, we should be able to safely
delete untagged images.

The exception are the various visualisation-base images which we do (for now)
use untagged, although this is being changed.