fix: visualisation-base lifecycle rule
This fixes the issue where "python" and "rv4" images were being expired from the
visualisation-base repository.

I incorrectly thought that if an image matches any of the patterns in the
tag_pattern_list in a rule, it would match the rule and stop processing.
However, it has to match _all_ of the patterns.

So to fix the issue, the one rule that matches "python","rv4" is split into two
rules.