fix: restore SageMaker access without access to other accounts
Recent changes incorrectly used specific resources in SageMaker actions in an
attempt to restrict access from tools to other AWS Accounts.

However, various SageMaker actions only support "*" as the resource, so access
has to be restricted using another mechanism - for example using a ConditionKey
on the Principal. This is what this change does.

It also slightly opens up the policy on the SageMaker VPC Endpoint to allow all
actions. This is done to keep the terraform/policy code minimal, and doesn't
actually increase what users(via their AWS role) can do, because they still
have an identity-based policy that restricts that.