refactor: refer to security group rather then VPC CIDR in rule
This makes it really clear what component needs to talk to the SageMaker
endpoints - the notebooks (i.e. tools), an dmean if we ever put anything else
in the notebooks VPC, it won't be able to to inadvertantly communicate with the
SageMaker endpoints.

(We maybe could do with doing similar things in other rules)