feat: Matchbox CodeBuild triggerable from GitHub merge or releases
This adds the ability for Matchbox's CodeBuild project to be triggered from
GitHub merge or release, configurable per environment.

This should allow the team to deploy Matchbox more easily in a more CD way

It does require the "CodeConnection" to GitHub to be setup manually for the
account Matchbox is deployed to. As far as I can tell, there is no way around
this being a manual step.