refactor: remove unused JupyterLab config
This removes unused config for the "JupyterLab" tool, which hasn't been used
for several years.

Note this is _not_ the same as "JupyterLab Python" which remains. "JupyterLab"
was the original tool in Data Workspace which has both R and Python in the same
tool.

This is done as a tidy-up in advance of adding CodeBuild projects for each tool