fix: codebuild compute types and apt-key port
Set 'codebuild_compute_type' to 'SMALL' for all tools but set it to 'MEDIUM' for theia
to fix codebuild project from failing due to running out of memory.

Created a security group rule to allow codebuild build builds to initiate
connections on port '13371' which allows apt-key to work, which is part of R
Studio's build.