Update of count type on boolean var (#339)
## What

The cloudwatch_subscription_filter variable is set as a boolean, and the
count statement for the matchbox aws_cloudwatch_log_subscription_filter
has been updated to reflect this.

## Checklist

<!--- The commands `git commit --amend`, `git rebase -i` and `git push
origin feat/my-change --force-with-lease` can be useful in acheiving the
following -->

* [x] Each commit in the PR is
[atomic](https://gds-way.cloudapps.digital/standards/source-code/working-with-git.html#atomic-commits).
In many cases a single commit in the PR is appropriate.
* [x] Each commit has a [good
message](https://gds-way.cloudapps.digital/standards/source-code/working-with-git.html#commit-messages),
with a body and not just a title, ideally adhering to the [Conventional
Commit Specification](https://www.conventionalcommits.org/en/v1.0.0/).
* [x] I have self-reviewed the PR - looking at the code changes and
descriptions of all its commits.

Co-authored-by: Sophie Glinton <sophieglinton@DIT000153.local>