refactor: inline Arango task definition
This inlines Arango's JSON ECS task definition

This is part of an ongoing tidy up of Terraform to make it easier to maintain.
And from experience, when changing the Terraform-level task definition, it is
extremely frequent to also change the JSON, and vice-versa. So it makes sense
for them to be together.