refactor: inline visualisation container definition (#376)
This inlines the container definition for (custom) visualisations.

Note that because "sentry_dsn" is no longer passed to container definition JSON
(because it was unused), the container definition JSON is no longer considered
sensitive, and so terraform apply treats this as an actual change.

This is part of the ongoing work to tidy up the Terraform