refactor: partially inline admin/frontend container definitions
This take a step to "inlining" the admin/frontend container defitions. It can
still be simplified I suspect, by a) converting the JSON string in Terraform
variable that is JSON-encoded, and b) putting all, or almost all, variables
into the Terraform variable rather than using a template* function. Leaving
this to a later change.

This is part of the ongoing tidy up of the Terraform