refactor: move visualisations into tools
This moves visualisations into tools, as a "sub feature".

Have changed my mind about this a few times, but they are very closely
assocaited with tools - run in the tools/notebooks VPC and developed from them,
and run from similar ecr images.

This is part of the tidy up of the Terraform