fix: finding tag of DNS rewrite proxy (#394)
This fixes finding the current tag of the DNS rewrite proxy, introduced since
the legacy DNS rewrite proxy was removed.

Without this it would revert to "latest", even if that's not the right tag. Not
sure why this didn't come up before - maybe the legacy service still existed
via the API - a lot of things in AWS are eventually consistent,