security: remove unnecessary QuickSight security group rule (#423)
This removes an unnecessary QuickSight security group rule, that theoretically
would allow the datasets database to initiate connections to QuickSight. I
strongly suspect this is already impossible/would acheive nothing already.

It also adds a comment to explain why there is an unusual security group rule
remaining that would not typically be needed for a client -> server connection.

This is done to be as secure as possible - standard minimisation of
privilege/swiss-cheese type of change. For the avoidance of doubt, I strongly
suspect even without this change it is already impossible for the database to
initial such a connection, certainly for a user to initiate such a connection,
and also impossible if such a connection were possible to use the connection
for anything.