security: remove GitLab ECS service access to ECR
This removes GitLab's security-group level access to ECR. Note that this
removes access from the service itself (i.e. from inside the running Docker
container), and not the EC2 instance that pulls the Docker image from ECR.

This is a belt-and-braces/swiss cheese change - doesn't close an actual known
security vulnerability for several reasons, but since GitLab shouldn't be able
to communicate with ECR, removing the security group rules that allows it to.