docs: clarify where security group rules should be created
This adds a policy to clarify where security group rules should be created.

Creating the first reusable module that needs security groups in
https://github.com/uktrade/data-workspace/pull/435, a module to create a
CodeBuild project that builds a Docker image and updates an ECS service, I
realised it wasn't exactly clear from the policy where they should go.

In this  case as well, the security group rules of a CodeBuild job are very
dependency on which VPC the CodeBuild job runs in, for example it needs rules
to open up to certain groups to communicate with the VPC endpoints that are
just in that VPC, so it needs the flexibility of being used with different
security groups.

Possibly the 'security group associations' feature could avoid needing this
flexibility, but this would make the infrastructure even more complicated