refactor: matchbox codebuild to use codebuild module (#437)
This refactors the Matchbox CodeBuild terraform to use our fairly new codebuild_ecs_deploy module. It includes a small fix to the module when using a non-default patch to the Dockerfile.

This is in keeping with the policy in infra/ to factor our modules, and in this case it keeps the amount of code/duplication down, especially around the boilerplate of creating a CodeBuild project.

Co-authored-by: Michal Charemza <michal@charemza.name>