infra: reduce gitlab deregistration delay
This reduces the deregistration delay for GitLab's target groups from the
default 5 minutes to 5 seconds.

This is to reduce the time it takes to deploy a new GitLab. This is all
acceptable from a user point of view because the deregistration delay is only
applicable when a GitLab instance is shutting down and in our setup this means
downtime anyway.