infra: disable (non functional) GitLab upload object storage (#454)
This disables the object storage for GitLab uploads.

The bucket is empty - it was never actually functional. However, having it
enabled now appears to break forks of projects with avatars. Forking such
projects results in an error mentioning that the bucket "uploads." does not
exist. This is right, the bucket is just , so it's odd (maybe a bug with
GitLab?)