infra: GitLab ECR lifecycle policy (#458)
This changes the GitLab ECR lifecycle policy to expire all but the most
recently pushed image.

This is the best we can come up with given the constraints of the ECR lifecycle
system. For example, we would base things on pull or push date, or make sure to
keep images that are running, but these are not possible.