infra: frontend CodeBuild deploys both webserver and celery
This makes it so the reusable CodeBuild module can deploy 0 or more ECS
services, and configures the frontend's CodeBuild project to deploy both the
webserver ("admin") and its Celery services ("admin-celery"). Before this
change, Celery was not deployable through CodeBuild (only manually or via a
separate Jenkins instance).

There are likely to be more changes after this one - specifically making it so
the Docker tag is customisable and using ecs-deploy for deployment to be
blocking (and so give better feedback).

We're moving away from Jenkins, and so need the components to be deployable
just via the Terraform here. This also makes the CodeBuild module more friendly
to be used for deployments that need to deploy 0 ECS services, i.e. tools.